fix big decimal conversions

+++ src/main/java/com/vega/protocol/model/LiquidityCommitmentOffset.java
@@ -5,11 +5,12 @@ import lombok.Data;
 import lombok.experimental.Accessors;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 
 @Data
 @Accessors(chain = true)
 public class LiquidityCommitmentOffset {
-    private Integer proportion;
+    private BigInteger proportion;
     private BigDecimal offset;
     private PeggedReference reference;
 }
